thermal: int340x: processor_thermal: Move mailbox code to common module
Mainline: commit b894685cb8e09305b3cb82bc39f2c4a3d154ce3a ("thermal: int340x: processor_thermal: Move mailbox code to common module")
from: v6.7-rc1    

The processor thermal mailbox is used for workload type request and
also in the processor thermal RFIM module. So, move the workload type
request code to its own module from the current processor thermal
mailbox module.

processor_thermal_mailbox.c contains only mailbox read/write related
source code. The source related to workload_types requests is moved to
a module processor_thermal_wt_req.c.

In addition
 - Rename PROC_THERMAL_FEATURE_MBOX to PROC_THERMAL_FEATURE_WT_REQ.
 - proc_thermal_mbox_add(), which adds workload type sysfs attribute group
   is renamed to proc_thermal_wt_req_add().
 - proc_thermal_mbox_remove() is renamed to proc_thermal_wt_req_remove().

While here, resolve check patch warnings for 100 columns for only modified
lines.

No functional changes are expected.

Intel-SIG: Upstream commit b894685cb8e0 ("thermal: int340x: processor_thermal: Move mailbox code to common module").

[ Baoli Zhang: amend commit log ]
